moved componentMediaQuery to baseComponent
